requirements.txt
python-telegram-bot
beautifulsoup4
requests
nltk
pandas
openai
python-dotenv
